added defensive program to reduce risk of none super users having ability to CRUD

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -70,8 +71,13 @@ def product_detail(request, product_id):
     return render(request, "products/product_detail.html", context)
 
 
+@login_required
 def add_product(request):
     '''A view to allow super users to add products'''
+    if not request.user.is_superuser:
+        messages.error(request, 'sorry only store owners can add products')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -92,8 +98,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def update_product(request, product_id):
     '''A view to allow super users to edit products'''
+    if not request.user.is_superuser:
+        messages.error(request, 'sorry only store owners can update products')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -116,8 +127,13 @@ def update_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     '''A view to allow super users to delete products'''
+    if not request.user.is_superuser:
+        messages.error(request, 'sorry only store owners can delete products')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f' {product.name} has been deleted')
